+++ app/client/src/reducers/entityReducers/metaReducer/index.ts
@@ -13,7 +13,10 @@ import {
 } from "@appsmith/constants/ReduxActionConstants";
 import produce from "immer";
 import type { EvalMetaUpdates } from "@appsmith/workers/common/DataTreeEvaluator/types";
-import { getMetaWidgetResetObj } from "./metaReducerUtils";
+import {
+  getMetaWidgetResetObj,
+  getNextMetaStateWithUpdates,
+} from "./metaReducerUtils";
 import type { WidgetEntityConfig } from "@appsmith/entities/DataTree/types";
 
 export type WidgetMetaState = Record<string, unknown>;
@@ -28,18 +31,7 @@ export const metaReducer = createReducer(initialState, {
       evalMetaUpdates: EvalMetaUpdates;
     }>,
   ) => {
-    const { evalMetaUpdates } = action.payload;
-
-    if (!evalMetaUpdates.length) return state;
-
-    // if metaObject is updated in dataTree we also update meta values, to keep meta state in sync.
-    const newMetaState = produce(state, (draftMetaState) => {
-      evalMetaUpdates.forEach(({ metaPropertyPath, value, widgetId }) => {
-        set(draftMetaState, [widgetId, ...metaPropertyPath], value);
-      });
-      return draftMetaState;
-    });
-    return newMetaState;
+    return getNextMetaStateWithUpdates(state, action);
   },
   [ReduxActionTypes.RESET_WIDGET_META_UPDATES]: (
     state: MetaState,
@@ -47,19 +39,7 @@ export const metaReducer = createReducer(initialState, {
       evalMetaUpdates: EvalMetaUpdates;
     }>,
   ) => {
-    const { evalMetaUpdates } = action.payload;
-
-    if (!evalMetaUpdates.length) return state;
-
-    // if metaObject is updated in dataTree we also update meta values, to keep meta state in sync.
-    const newMetaState = produce(state, (draftMetaState) => {
-      evalMetaUpdates.forEach(({ metaPropertyPath, value, widgetId }) => {
-        set(draftMetaState, [widgetId, ...metaPropertyPath], value);
-      });
-      return draftMetaState;
-    });
-
-    return newMetaState;
+    return getNextMetaStateWithUpdates(state, action);
   },
   [ReduxActionTypes.SET_META_PROP]: (
     state: MetaState,
